// generated/src/app/api-models/linbs-contract-search.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class LinbsContractSearch {
  lineOfBusiness: string;
  contractSearchType: string;
  contractSearchOrder: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}